requirements.txt
appdirs>=1.4.4
asgiref>=3.2.10
beautifulsoup4==4.9.3
better-exceptions>=0.2.2
django-bootstrap4>=2.3.1
cached-property>=1.5.2
certifi>=2020.6.20
cfenv>=0.5.3
cffi>=1.14.3
chardet>=3.0.4
click>=7.1.2
cloudflare>=2.8.15
cmake>=3.18.2.post1
colorama>=0.4.4
cryptography>=3.1.1
cycler>=0.10.0
django-session-csrf>=0.7.1
django-bootstrap4>=2.3.1
djangorestframework>=3.12.4
decorator>=4.4.2
Django>=3.1.2
django-filter>=2.4.0
dj-database-url>=0.5.0
future>=0.18.2
furl>=2.1.0
https-everywhere>=0.2.2
idna>=2.10
imageio>=2.9.0
jsonlines>=2.0.0
kiwisolver>=1.2.0
logging-helper>=1.9.0
logging-tree>=1.8.1
matplotlib>=3.3.2
mouseinfo>=0.1.3
mindsphere_core-1.0.2-py3-none-any.whl
networkx>=2.5
numpy>=1.19.2
opencv-python>=4.4.0.44
orderedmultidict>=1.0.1
oauthlib>=3.1.0
pandas>=1.1.2
Pillow>=6.2.1
pycparser>=2.20
PyJWT>=1.7.1
pyrect>=0.1.4
psycopg2>=2.8.6
PyAutoGUI>=0.9.52
pygetwindow>=0.0.5
pymsgbox>=1.0.9
pyOpenSSL>=19.1.0
pyparsing>=2.4.7
pyperclip>=1.8.1
pyscreeze>=0.1.26
PyTweening>=1.0.1
python-dateutil>=2.8.1
pytz>=2020.1
PyWavelets>=1.1.1
redis>=3.5.3
requests>=2.24.0
scikit-image>=0.17.2
scipy>=1.5.2
six>=1.15.0
soupsieve>=2.2
sqlparse>=0.4.1
tifffile>=2020.9.29
timeseries-3.4.1-py3-none-any.whl
urllib3>=1.25.11
WhiteNoise>=5.2.0
xhtml2pdf>=0.2.5
iottsbulk-3.5.0-py3-none-any.whl